Refactor completion item logic into separate modules
Extracted functions related to completion items and ALPS profile parsing into their own modules. This improves code organization and readability, and prepares the project for easier future modifications.
